mejora en vista verPostulantes

+++ database/migrations/2024_12_21_000005_create_perfiles_desarrollador_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('CBU');
             $table->string('telefono');
             $table->text('descripcion_sobre_mi')->nullable();
-            $table->integer('promedio_calificacion')->nullable();
+            $table->float('promedio_calificacion')->nullable();
             $table->foreignId('id_usuario');
         });
     }
